requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
chardet==3.0.4
Django==2.2
djangorestframework==3.11.0
idna==2.8
multidict==4.7.4
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
yarl==1.4.2
